Join appears to work (requires unit tests).
Fix race condition with multiple sources; FlowElements now close only when
all upstream elements notify them of closure.
HashJoinElement now calculates join time ranges using different intervals
for left-on-right and right-on-left.
WindowedHashMap supports open and closed ends for both sides of the interval.
WindowedHashMap supports oldestTimestamp() method.
HashJoinElement now calculates expiry in the insertMap based on the range
perspective of the oldest timestamp still in the joinMap, not based on
the time of the element in the insertMap.