Add AvroEventParser.
Added AvroEventParser to read input events which are in binary-encoded Avro
form (as, for example, are the output events from an OutputElement via Flume).
Registered AvroEventParser with FormatSpec.
Got RtsqlSink using FormatSpec to instantiate the EventParser.
Implemented required API changes for this in several other classes
(RtsqlSink knew the stream's name, but not its StreamSymbol).

Bugfix: FlowId is now Comparable (required for use in a TreeMap).