RTSQL-14. 2nd query on foreign Flume node fails.
Fixed issue where a second query could not be executed against
the same foreign Flume logical node in a reasonable amount of time.
Flume may reconfigure logical nodes relatively quickly in the master,
but the actual target node won't be replaced until another event
prompts it. If there's an agent--collector connection in the middle,
this basically means waiting until the agent TCP timeouts trying to
write to the old collector.

We now no longer use FanOutSink to handle multiple RtsqlSink instances
in the embedded Flume environment. This change introduces RtsqlMultiSink
which can handle local (thread-safe) modifications to the fanout, and
can be accessed through a process-wide cache by the other actors who
need to find a specific MultiSink.